+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `PropertyFactory` must implement new method `makeCustom`.
 - Support for method's parameters/return types retrieved from phpDoc.
+- Add a test generator for Laravel Broadcasting channel.
 
 **Changed**